+++ mercury/src/main/java/org/broadinstitute/gpinformatics/mercury/entity/sample/SampleInstanceEntity.java
@@ -57,7 +57,7 @@ public class SampleInstanceEntity {
 
     private String experiment;
 
-    private Long readLength;
+    private Integer readLength;
 
     public void removeSubTasks() {
         sampleInstanceEntityTsks.clear();
@@ -84,9 +84,9 @@ public class SampleInstanceEntity {
        return subTask;
     }
 
-    public Long getReadLength() { return readLength; }
+    public Integer getReadLength() { return readLength; }
 
-    public void setReadLength(Long readLength) { this.readLength = readLength;  }
+    public void setReadLength(Integer readLength) { this.readLength = readLength;  }
 
     public MercurySample getRootSample() {  return rootSample;  }
